Kamal (#42)
* #1

* complete AddToken Feature

* AddToken Completed

* Posts

* changes in lead

* sales updated

* 20/08/2025

* 02/09/2025

* 04/09/2025

+++ src/main/java/com/realestate/main/controller/AgentController.java
@@ -168,7 +168,7 @@ return new ResponseEntity<TokenDto>(payment,HttpStatus.OK);
 	
 	@PreAuthorize("hasRole('Agent')")
 	@PutMapping("/payment")
-	public ResponseEntity<TokenDto> makepayment(@RequestParam int  tokenId,@RequestParam double finalAmount,@RequestParam String finalStatus) throws TokenNotFoundException{
+	public ResponseEntity<TokenDto> makepayment(@RequestParam int  tokenId,@RequestParam long finalAmount,@RequestParam String finalStatus) throws TokenNotFoundException{
 		TokenDto payment = agentService.makePayment(tokenId, finalAmount, finalStatus);
 		return new ResponseEntity<TokenDto>(payment,HttpStatus.OK);
